requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy>=2.0.38
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic>=2.5.3
pydantic-settings==2.1.0
requests>=2.32.3
gmssl>=3.2.2
alembic==1.13.1
apscheduler==3.10.4
email-validator==2.1.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
aiohttp==3.9.3
cloudscraper==1.2.71